Add C99 compliant snprintf to libcompat
Many systems do not provide a C99 compliant snprintf implementation.
To guarantee that the snprintf used by parts of check is correct,
a version is added to libcompat. The system's snprintf is checked with
the configure script. If the system's snprintf is not up to snuff,
the libcompat version is used.

This version of snprintf is from Holger Weiß and Patrick Powell.
http://www.jhweiss.de/software/snprintf.html

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@687 64e312b2-a51f-0410-8e61-82d0ca0eb02a