Always try alternative to tmpfile() if necessary
If tmpfile fails on a platform, there is no harm in trying the
alternative.

Previously the alternative was only attempted for Windows. However,
checking for WIN32 was not correct; instead _WIN32 should be used.
When the proper check for Windows was attempted, OSX would fail
to run its tests. However, removing the conditional and renaming
_tempnam to tempnam (which is available in MinGW, OSX, and GNU/Linux)
works for all three platforms.

Correctly check for Windows environment

The correct way to check for compiling for Windows is to check for
_WIN32 instead of WIN32.

Patch submitted by username bross on Sourceforge, patch#48

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@756 64e312b2-a51f-0410-8e61-82d0ca0eb02a