test_xml_output: remove check for -1 in durations
This test is only valid on platforms that have a functional
clock_gettime() and do not have a good replacement in libcompat.
If there is no good replacement, then all durations will be 0,
even during failures. Instead of forcing the duration to be
negative during a failure artificially, we remove the requirement.

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@801 64e312b2-a51f-0410-8e61-82d0ca0eb02a