Remove -Werror from automake
On newer versions of automake, doc/Makefile.am causes a warning:

It appears this file (or files included by it) are triggering
an undocumented, soon-to-be-removed automake hack.
Future automake versions will no longer place in the builddir
(rather than in the srcdir) the generated '.info' files that
appear to be cleaned, by e.g. being listed in CLEANFILES or
DISTCLEANFILES. If you want your '.info' files to be placed in the
builddir rather than in the srcdir, you have to use the
shiny new 'info-in-builddir' automake option.

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@824 64e312b2-a51f-0410-8e61-82d0ca0eb02a