configure.ac: do not compile with -ansi for now
There is currently a bug on the MinGW/MinGW-w64 platforms:
   sourceforge.net/p/mingw/bugs/2024
where compiling with -ansi will not define the off64_t type,
but including unistd.h will require it to be defined else
a compile error will result.

Attempting to define a replacement in config.h also fails; config.h
will work for check's code, but the code that libtool produces
then fails, meaning unit tests fail to compile.

When the MinGW bug is resolved, we can start compiling with -ansi again.

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@846 64e312b2-a51f-0410-8e61-82d0ca0eb02a