add support for logging in TAP format
Logging in the Test Anything Protcol (TAP) format is useful
in general, but is especially useful on platforms where
fork() is unavailable. The TAP format requires that the
planned number of tests to run be specified, either at the
start or end of a test run. If the number of tests run does not
match the plan, the test run is marked as failed. As running
without fork() leaves the unit testing program open to failure
if a test aborts/segfaults/etc, validating the TAP output is
useful.

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@873 64e312b2-a51f-0410-8e61-82d0ca0eb02a