modify how HAVE_FORK is determined in code
autotools will define HAVE_FORK only if fork is available.
cmake can do the same thing on Linux, however on Windows
the only way to get HAVE_FORK to appear is to make it always
appear and contain a value. To support both autotools and
cmake, change the compile time check slightly to check that
HAVE_FORK is 1.

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@956 64e312b2-a51f-0410-8e61-82d0ca0eb02a