libcompat: do not offer to replace setenv
if setenv is missing on the system, the only thing that is
affected are unit tests. Check itself does not need setenv.
In this case, the affected unit tests are not compiled in.

This change removed the stub setenv replacement which was
never called, even when a system did not provide it.

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@984 64e312b2-a51f-0410-8e61-82d0ca0eb02a