libcompat: remove putenv replacement
the putenv replacement did nothing except call abort().
There is no platform (yet) which did not have putenv()
but did have fork(). For this reason, the assert()
has yet to be hit, meaning that the stub itself is
pointless.

If in the future a platform is found to have fork() but
not putenv(), then the related tests need to be disabled
at compile time using something like
  #if defined(HAVE_DECL_PUTENV)
along with a related AC_CHECK_DECLS check.

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@986 64e312b2-a51f-0410-8e61-82d0ca0eb02a