libcompat: remove unsetenv replacement
no system to date has been found that provides setenv
but not unsetenv. If setenv is unavailable, all related
tests are disabled at compile time. This stub replacement
does not serve any purpose. If in the future a system
with unsetenv is found that does provide setenv, then
the unit tests will need to disable all tests that
require unsetenv.

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@987 64e312b2-a51f-0410-8e61-82d0ca0eb02a